-get_e some problem

+++ ft_get_float.c
@@ -6,7 +6,7 @@
 /*   By: trofidal <trofidal@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/06 21:28:35 by trofidal          #+#    #+#             */
-/*   Updated: 2021/03/07 11:22:59 by trofidal         ###   ########.fr       */
+/*   Updated: 2021/03/08 08:36:04 by trofidal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char	*ft_get_float(char *flags, va_list args)
 	if (ft_strlen(flags) == 0)
 	{
 		va_arg = va_arg(args, double);
-		flt = ft_putnbr_float((long double) va_arg);
+		flt = ft_putnbr_float(va_arg);
 		return (flt);
 	}
 	return (NULL);
